[WIP] use type alias

+++ dart_native/lib/src/android/foundation/wrapperclass/long.dart
@@ -17,7 +17,7 @@ class Long extends JSubclass<int> {
 /// New native 'Long'.
 Pointer<Void> _new(dynamic value, String clsName) {
   if (value is int) {
-    JObject object = JObject.parameterConstructor(clsName, [long(value)]);
+    JObject object = JObject.parameterConstructor(clsName, [value]);
     return object.pointer.cast<Void>();
   } else {
     throw 'Invalid param when initializing Long.';
